requirements.txt
alembic==1.5.4
aniso8601==9.0.1
astroid==2.4.2
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
google-trans-new==1.1.9
graphql-core==2.3.2
graphql-server-core==1.2.0
idna==3.2
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
jwt==1.2.0
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
promise==2.3
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.4.2
pylint==2.6.0
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.4
requests==2.26.0
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.6
Werkzeug==1.0.1
wrapt==1.12.1
